# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: NovaCron VM Management API
  description: Distributed virtual machine management platform API
  version: 1.0.0
  contact:
    name: NovaCron Team
    email: support@novacron.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.novacron.io/v1
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /migrations/{migration_id}:
    get:
      tags:
        - Migration
      summary: Get migration status
      parameters:
        - name: migration_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Migration status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Migration'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Migration:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vm_id:
          type: string
          format: uuid
        source_host:
          type: string
          format: uuid
        target_host:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, preparing, transferring, completing, completed, failed, cancelled]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        error_message:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...